+++ src/index.js
@@ -8,6 +8,10 @@ import { createBrotliCompress, createBrotliDecompress } from 'zlib';
 import crypto from 'crypto';
 import { goUp } from './nwd/up.js';
 import { changeDirectory } from './nwd/changeDir.js';
+import {listDirectory} from './nwd/ls.js'
+import { readContnet } from './fileManipulations/cat.js';
+import { createFile } from './fileManipulations/createFile.js';
+import { renameFile } from './fileManipulations/renameFile.js';
 
 const args = process.argv.slice(2);
 const usernameArg = args.find(arg => arg.startsWith('--username='));
@@ -37,7 +41,7 @@ rl.on('line', async (input) => {
         listDirectory();
         break;
       case 'cat':
-        readFile(args[0]);
+        readContnet(args[0]);
         break;
       case 'add':
         createFile(args[0]);
@@ -88,103 +92,4 @@ function exitProgram() {
 
 function printCurrentDirectory() {
   console.log(`You are currently in ${process.cwd()}`);
-}
-
-function listDirectory() {
-  const items = fs.readdirSync(currentDir);
-  const directories = items.filter(item => fs.lstatSync(path.join(currentDir, item)).isDirectory()).sort();
-  const files = items.filter(item => fs.lstatSync(path.join(currentDir, item)).isFile()).sort();
-  directories.forEach(dir => console.log(`DIR: ${dir}`));
-  files.forEach(file => console.log(`FILE: ${file}`));
-}
-
-function readFile(filePath) {
-  const fullPath = path.isAbsolute(filePath) ? filePath : path.join(currentDir, filePath);
-  const stream = fs.createReadStream(fullPath, 'utf-8');
-  stream.pipe(process.stdout);
-}
-
-function createFile(fileName) {
-  const filePath = path.join(currentDir, fileName);
-  fs.writeFileSync(filePath, '', 'utf-8');
-}
-
-function renameFile(oldPath, newFileName) {
-  const fullOldPath = path.isAbsolute(oldPath) ? oldPath : path.join(currentDir, oldPath);
-  const newPath = path.join(path.dirname(fullOldPath), newFileName);
-  fs.renameSync(fullOldPath, newPath);
-}
-
-function copyFile(srcPath, destPath) {
-  const fullSrcPath = path.isAbsolute(srcPath) ? srcPath : path.join(currentDir, srcPath);
-  const fullDestPath = path.isAbsolute(destPath) ? destPath : path.join(currentDir, destPath);
-  const readStream = fs.createReadStream(fullSrcPath);
-  const writeStream = fs.createWriteStream(fullDestPath);
-  readStream.pipe(writeStream);
-}
-
-function moveFile(srcPath, destPath) {
-  copyFile(srcPath, destPath);
-  deleteFile(srcPath);
-}
-
-function deleteFile(filePath) {
-  const fullPath = path.isAbsolute(filePath) ? filePath : path.join(currentDir, filePath);
-  fs.unlinkSync(fullPath);
-}
-
-function handleOsCommand(option) {
-  switch (option) {
-    case '--EOL':
-      console.log(JSON.stringify(os.EOL));
-      break;
-    case '--cpus':
-      const cpus = os.cpus();
-      console.log(`Total CPUs: ${cpus.length}`);
-      cpus.forEach((cpu, index) => {
-        console.log(`CPU ${index + 1}: ${cpu.model}, ${cpu.speed / 1000} GHz`);
-      });
-      break;
-    case '--homedir':
-      console.log(os.homedir());
-      break;
-    case '--username':
-      console.log(os.userInfo().username);
-      break;
-    case '--architecture':
-      console.log(process.arch);
-      break;
-    default:
-      console.log('Invalid input');
-  }
-}
-
-function calculateHash(filePath) {
-  const fullPath = path.isAbsolute(filePath) ? filePath : path.join(currentDir, filePath);
-  const hash = crypto.createHash('sha256');
-  const stream = fs.createReadStream(fullPath);
-  stream.on('data', (chunk) => hash.update(chunk));
-  stream.on('end', () => console.log(hash.digest('hex')));
-}
-
-function compressFile(srcPath, destPath) {
-  const fullSrcPath = path.isAbsolute(srcPath) ? srcPath : path.join(currentDir, srcPath);
-  const fullDestPath = path.isAbsolute(destPath) ? destPath : path.join(currentDir, destPath);
-  const readStream = fs.createReadStream(fullSrcPath);
-  const writeStream = fs.createWriteStream(fullDestPath);
-  const brotli = createBrotliCompress();
-  pipeline(readStream, brotli, writeStream, (err) => {
-    if (err) console.log('Operation failed');
-  });
-}
-
-function decompressFile(srcPath, destPath) {
-  const fullSrcPath = path.isAbsolute(srcPath) ? srcPath : path.join(currentDir, srcPath);
-  const fullDestPath = path.isAbsolute(destPath) ? destPath : path.join(currentDir, destPath);
-  const readStream = fs.createReadStream(fullSrcPath);
-  const writeStream = fs.createWriteStream(fullDestPath);
-  const brotli = createBrotliDecompress();
-  pipeline(readStream, brotli, writeStream, (err) => {
-    if (err) console.log('Operation failed');
-  });
 }
